refactor: seperate user, post from main
