+++ src/features/post/ui/PostSortFilter.tsx
@@ -1,0 +1,31 @@
+import useParam from "../../../shared/model/useParam"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui"
+
+export const PostSortFilter = () => {
+  const {sortBy, sortOrder, updateParam} = useParam()
+
+  return (
+    <>
+      <Select value={sortBy} onValueChange={(value) => updateParam("sortBy", value)}>
+        <SelectTrigger className="w-[180px]">
+          <SelectValue placeholder="정렬 기준" />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectItem value="none">없음</SelectItem>
+          <SelectItem value="id">ID</SelectItem>
+          <SelectItem value="title">제목</SelectItem>
+          <SelectItem value="reactions">반응</SelectItem>
+        </SelectContent>
+      </Select>
+      <Select value={sortOrder} onValueChange={(value) => updateParam("sortOrder", value)}>
+        <SelectTrigger className="w-[180px]">
+          <SelectValue placeholder="정렬 순서" />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectItem value="asc">오름차순</SelectItem>
+          <SelectItem value="desc">내림차순</SelectItem>
+        </SelectContent>
+      </Select>
+    </>
+  )
+}
